oxxxymiron/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django
django-extensions==3.1.5
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.19.2
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
